Restructure data folder
Changed data folder structure from data/<data type>/constants.py and memory.py to data/constants and data/memory_addrs with files of each data type as necessary. Done bcause many data types (statuses/types/etc.) are small and don't have memory mappings in the way larger structures like Pokemon do, and making an entire directory with a single .py file containing < 20 lines felt silly.
