+++ pyboy/plugins/game_wrapper_pokemon_gen1/data/constants/locations.py
