+++ pyboy/plugins/game_wrapper_pokemon_gen1/data/constants/pokemon.py
