+++ pyboy/plugins/game_wrapper_pokemon_gen1/data/memory_addrs/pokemon.py
